added versioning to crypto package

+++ implementation/tfdacmacs/crypto/src/main/java/de/tuberlin/tfdacmacs/crypto/pairing/data/keys/TwoFactorUpdateKey.java
@@ -4,7 +4,7 @@ import it.unisa.dia.gas.jpbc.Element;
 import lombok.NonNull;
 
 public class TwoFactorUpdateKey extends UserUpdateKey {
-    public TwoFactorUpdateKey(@NonNull String userId, @NonNull Element updateKey, int targetVersion) {
+    public TwoFactorUpdateKey(@NonNull String userId, @NonNull Element updateKey, long targetVersion) {
         super(userId, updateKey, targetVersion);
     }
 }
