+++ implementation/tfdacmacs/crypto/src/main/java/de/tuberlin/tfdacmacs/crypto/pairing/data/keys/UserUpdateKey.java
@@ -9,7 +9,7 @@ public abstract class UserUpdateKey extends UpdateKey {
 
     private final String userId;
 
-    public UserUpdateKey(@NonNull String userId, @NonNull Element updateKey, int targetVersion) {
+    public UserUpdateKey(@NonNull String userId, @NonNull Element updateKey, long targetVersion) {
         super(updateKey, targetVersion);
          this.userId = userId;
     }
